+++ service/monitor/listener_test.go
@@ -76,11 +76,11 @@ func testNewMetricsListenerUpdate(t *testing.T) {
 		now = float64(time.Now().Unix())
 
 		p = xmetricstest.NewProvider(nil, service.Metrics).
-			Expect(service.UpdateCount, service.ServiceLabel, "talaria")(xmetricstest.Value(1.0)).
-			Expect(service.LastUpdateTimestamp, service.ServiceLabel, "talaria")(xmetricstest.Minimum(now)).
-			Expect(service.ErrorCount, service.ServiceLabel, "talaria")(xmetricstest.Value(0.0)).
-			Expect(service.LastErrorTimestamp, service.ServiceLabel, "talaria")(xmetricstest.Value(0.0)).
-			Expect(service.InstanceCount, service.ServiceLabel, "talaria")(xmetricstest.Value(2.0))
+			Expect(service.UpdateCount, service.ServiceLabel, "talaria", service.EventKeyLabel, "test")(xmetricstest.Value(1.0)).
+			Expect(service.LastUpdateTimestamp, service.ServiceLabel, "talaria", service.EventKeyLabel, "test")(xmetricstest.Minimum(now)).
+			Expect(service.ErrorCount, service.ServiceLabel, "talaria", service.EventKeyLabel, "test")(xmetricstest.Value(0.0)).
+			Expect(service.LastErrorTimestamp, service.ServiceLabel, "talaria", service.EventKeyLabel, "test")(xmetricstest.Value(0.0)).
+			Expect(service.InstanceCount, service.ServiceLabel, "talaria", service.EventKeyLabel, "test")(xmetricstest.Value(2.0))
 		l = NewMetricsListener(p)
 	)
 
@@ -93,11 +93,11 @@ func testNewMetricsListenerError(t *testing.T) {
 		now = float64(time.Now().Unix())
 
 		p = xmetricstest.NewProvider(nil, service.Metrics).
-			Expect(service.UpdateCount, service.ServiceLabel, "scytale")(xmetricstest.Value(0.0)).
-			Expect(service.LastUpdateTimestamp, service.ServiceLabel, "scytale")(xmetricstest.Value(0.0)).
-			Expect(service.ErrorCount, service.ServiceLabel, "scytale")(xmetricstest.Value(1.0)).
-			Expect(service.LastErrorTimestamp, service.ServiceLabel, "scytale")(xmetricstest.Minimum(now)).
-			Expect(service.InstanceCount, service.ServiceLabel, "scytale")(xmetricstest.Value(0.0))
+			Expect(service.UpdateCount, service.ServiceLabel, "scytale", service.EventKeyLabel, "test")(xmetricstest.Value(0.0)).
+			Expect(service.LastUpdateTimestamp, service.ServiceLabel, "scytale", service.EventKeyLabel, "test")(xmetricstest.Value(0.0)).
+			Expect(service.ErrorCount, service.ServiceLabel, "scytale", service.EventKeyLabel, "test")(xmetricstest.Value(1.0)).
+			Expect(service.LastErrorTimestamp, service.ServiceLabel, "scytale", service.EventKeyLabel, "test")(xmetricstest.Minimum(now)).
+			Expect(service.InstanceCount, service.ServiceLabel, "scytale", service.EventKeyLabel, "test")(xmetricstest.Value(0.0))
 		l = NewMetricsListener(p)
 	)